refactor: Usage 및 User 관련 패키지 경로 및 메서드 수정
1. UsageResponse, UserInfoResponse, UserCreateRequest 패키지를 `adapter.in.web`으로 이동한다.
2. UsageResponse의 `from` 메서드를 UsageResult를 기반으로 수정한다.
3. UserInfoResponse의 `from` 메서드를 UserInfoResult를 기반으로 수정한다.
4. UserCreateRequest에 `toCommand` 메서드를 추가하여 UserCreateCommand로 변환하도록 수정한다.

Resolves #47
