+++ src/main/java/com/personalproject/llmmanager/users/adapter/in/web/response/UsageResponse.java
@@ -1,0 +1,29 @@
+package com.personalproject.llmmanager.users.adapter.in.web.response;
+
+import com.personalproject.llmmanager.users.application.dtos.result.UsageResult;
+import com.personalproject.llmmanager.users.domain.policy.ModelUsageDetails;
+import com.personalproject.llmmanager.users.domain.policy.UsageSummary;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public record UsageResponse (
+        String plan,
+        long quota,
+        long usedTokens,
+        long remainingTokens,
+        BigDecimal totalPrice,
+        List<ModelUsageDetails> models
+) {
+    static public UsageResponse from(UsageResult result) {
+        return new UsageResponse(
+                result.plan(),
+                result.quota(),
+                result.usedTokens(),
+                result.remainingTokens(),
+                result.totalPrice(),
+                result.models()
+        );
+    }
+}
